+++ plugin/src/test/scala/com/stratio/cassandra/lucene/IndexOptionsTest.scala
@@ -16,7 +16,7 @@
 package com.stratio.cassandra.lucene
 
 import com.stratio.cassandra.lucene.IndexOptions._
-import com.stratio.cassandra.lucene.partitioning.{PartitionerOnNone, PartitionerOnToken, PartitionerOnVirtualNodes$}
+import com.stratio.cassandra.lucene.partitioning.{PartitionerOnNone, PartitionerOnToken}
 import org.junit.runner.RunWith
 import org.scalatest.junit.JUnitRunner