+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 3.0.10.3 (Upcoming)
 
 * Fix deletion of ranges of rows in a partition
-* Add virtual nodes token ranges based index partitioner
+* Add virtual node based partitioner
 
 ## 3.0.10.2 (January 10, 2017)